requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
requests
django==2.2.6
pytest==6.2.2
pytest-pythonpath==0.7.3
pytest-django==4.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-filter==2.4.0
psycopg2-binary==2.8.5
python-dotenv==0.19.1
gunicorn==20.1.0